+++ packages/core/Example/vite-env.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="vite/client" />
 
 interface ImportMetaEnv {
-  readonly VITE_HEEDKIT_PROJECT_KEY?: string;
+  readonly VITE_HEEDKIT_WORKSPACE_KEY?: string;
   readonly VITE_HEEDKIT_API_URL?: string;
 }